refactor(tui): extract HomeScreen from app.py; slim router to 109 lines (closes #51)
- app.py: stripped to CanvasTUI App router (~109 lines); all state/logic moves
  to HomeScreen; CLI overrides forwarded via constructor overrides dict
- screens/home.py: ~1540 lines holding all state, CSS, BINDINGS, action_*
  methods, render helpers, background thread lifecycle; child screens receive
  HomeScreen as owner_app (has same attribute surface)
- screens/__init__.py: exports HomeScreen, BaseScreen, RMPScreen alongside
  existing screen exports
- screens/base.py: BaseScreen ABC with show_help_overlay() (written in Task 1)
- screens/rmp.py: RMPScreen three-state screen (search/results/details) —
  full implementation ships in 08-02; available now for __init__ export
