+++ src/canvas_tui/screens/__init__.py
@@ -3,6 +3,7 @@
 
 from .analytics import AnalyticsScreen
 from .announcements import AnnouncementDetailScreen, AnnouncementsScreen
+from .base import BaseScreen
 from .course_overview import CourseOverviewScreen
 from .courses import CourseManagerScreen
 from .dashboard import DashboardScreen
@@ -10,7 +11,9 @@ from .details import DetailsScreen
 from .files import FileManagerScreen
 from .grades import GradesScreen
 from .help import HelpScreen
+from .home import HomeScreen
 from .modals import ConfirmPath, InputPrompt, LoadingScreen
+from .rmp import RMPScreen
 from .settings import SettingsScreen
 from .syllabi import SyllabiScreen
 from .weekview import WeekViewScreen
@@ -19,6 +22,7 @@ __all__ = [
     "AnalyticsScreen",
     "AnnouncementDetailScreen",
     "AnnouncementsScreen",
+    "BaseScreen",
     "ConfirmPath",
     "CourseManagerScreen",
     "CourseOverviewScreen",
@@ -27,8 +31,10 @@ __all__ = [
     "FileManagerScreen",
     "GradesScreen",
     "HelpScreen",
+    "HomeScreen",
     "InputPrompt",
     "LoadingScreen",
+    "RMPScreen",
     "SettingsScreen",
     "SyllabiScreen",
     "WeekViewScreen",
